Factor out Command into its own files and update its tests.
